Refactor authentication and user management:
- Moved login and registration endpoints to `AuthController`.
- Updated `CaseController` annotations to use `@Controller` and added `@ResponseBody` to applicable methods.
- Removed unused methods from `UserController`.
- Fixed Thymeleaf syntax errors in templates.
- Updated form actions in templates to align with new endpoint mappings.
- Removed form-based login configuration from `SecurityConfig`.
