+++ src/main/java/org/example/untitled/usercase/controller/CaseController.java
@@ -9,12 +9,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@RestController
+@Controller
 @RequestMapping("/tickets")
 public class CaseController {
 
@@ -38,12 +39,14 @@ public class CaseController {
     }
 
     @GetMapping
+    @ResponseBody
     @PreAuthorize("hasAnyRole('HANDLER', 'ADMIN')")
     public ResponseEntity<List<CaseEntityDto>> getAllTickets() {
         return ResponseEntity.ok(caseService.getAllTickets());
     }
 
     @PutMapping("/{id}/status")
+    @ResponseBody
     @PreAuthorize("hasAnyRole('HANDLER', 'ADMIN')")
     public ResponseEntity<CaseEntityDto> updateStatus(
             @PathVariable Long id, @RequestParam CaseStatus status) {
@@ -51,6 +54,7 @@ public class CaseController {
     }
 
     @PutMapping("/{id}/assign")
+    @ResponseBody
     @PreAuthorize("hasAnyRole('HANDLER', 'ADMIN')")
     public ResponseEntity<CaseEntityDto> assignToSelf(
             @PathVariable Long id, @AuthenticationPrincipal UserDetails userDetails) {
